Changed JRE compliance from 7 to 6 + code cleanup
Cleaned up the code using the 'Inspect Code' tool in IntelliJ.
Changed the JRE compliance from 7 to 6 to make sure Excellence runs on
computers still running Java 6.
Added error message when a FunctionCell doesn't contain a Function, but
=B5 for instance.
Added ability to reference another cell directly by using, for example,
=A6 to reference the Cell at position A6. (Also added a check to prevent
StackOverflow Errors if a recursive reference is entered)

+++ src/main/nl/tudelft/excellence/functions/MEDIAN.java
@@ -33,8 +33,8 @@ public class MEDIAN extends NumberFunction{
 		Arrays.sort(input);
 		double res = 0;
 		if(input.length %2==0){
-			int a = (int) input.length/2-1;
-			int b = (int) input.length/2;
+			int a = input.length/2-1;
+			int b = input.length/2;
 			res = (input[a]+input[b])/2;
 		}
 		else if(input.length %2 != 0){
